Test errors fixed
